// src/app/core/api/generated/model/updateUserProfileRequestDto.ts
/**
 * Reusable Identity API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateUserProfileRequestDto { 
    displayName: string;
    firstName?: string | null;
    lastName?: string | null;
    phoneNumber?: string | null;
    avatarUrl?: string | null;
    timeZone?: string | null;
    preferredLanguage?: string | null;
    bio?: string | null;
}